+++ CHANGELOG.rst
@@ -4,9 +4,9 @@ Changelog
 
 Notable changes to this project will be documented in this file.
 
-********
-Upcoming
-********
+*******************
+v2.9.0 (2025-10-14)
+*******************
 
 Added
 =====